Interface fix and stricter qualification of signedness in tests.

+++ test/Utilities/NestedIterator_test.cc
@@ -88,8 +88,8 @@ void RunVectorVectorTest() {
     ++iElem;
   } // while
   
-  BOOST_CHECK_EQUAL(expected, NElements);
-  BOOST_CHECK_EQUAL(nMismatches, 0);
+  BOOST_CHECK_EQUAL((unsigned int) expected, NElements);
+  BOOST_CHECK_EQUAL(nMismatches, 0U);
 } // RunVectorVectorTest()
 
 
@@ -140,8 +140,8 @@ void RunVectorMapTest() {
     ++iElem;
   } // while
   
-  BOOST_CHECK_EQUAL(expected, NElements);
-  BOOST_CHECK_EQUAL(nMismatches, 0);
+  BOOST_CHECK_EQUAL((unsigned int) expected, NElements);
+  BOOST_CHECK_EQUAL(nMismatches, 0U);
 } // RunVectorMapTest()
